+++ src/video_core/CMakeLists.txt
@@ -42,6 +42,8 @@ add_library(video_core STATIC
     renderer_opengl/gl_resource_manager.h
     renderer_opengl/gl_shader_cache.cpp
     renderer_opengl/gl_shader_cache.h
+    renderer_opengl/gl_shader_decompiler.cpp
+    renderer_opengl/gl_shader_decompiler.h
     renderer_opengl/gl_shader_gen.cpp
     renderer_opengl/gl_shader_gen.h
     renderer_opengl/gl_shader_manager.cpp
@@ -83,8 +85,6 @@ add_library(video_core STATIC
     shader/decode/xmad.cpp
     shader/decode/other.cpp
     shader/decode.cpp
-    shader/glsl_decompiler.cpp
-    shader/glsl_decompiler.h
     shader/shader_ir.cpp
     shader/shader_ir.h
     surface.cpp
