+++ src/video_core/renderer_opengl/gl_shader_cache.cpp
@@ -10,9 +10,9 @@
 #include "video_core/engines/maxwell_3d.h"
 #include "video_core/renderer_opengl/gl_rasterizer.h"
 #include "video_core/renderer_opengl/gl_shader_cache.h"
+#include "video_core/renderer_opengl/gl_shader_decompiler.h"
 #include "video_core/renderer_opengl/gl_shader_manager.h"
 #include "video_core/renderer_opengl/utils.h"
-#include "video_core/shader/glsl_decompiler.h"
 #include "video_core/shader/shader_ir.h"
 
 namespace OpenGL {
