+++ src/video_core/renderer_opengl/gl_shader_gen.cpp
@@ -5,8 +5,8 @@
 #include <fmt/format.h>
 #include "common/assert.h"
 #include "video_core/engines/maxwell_3d.h"
+#include "video_core/renderer_opengl/gl_shader_decompiler.h"
 #include "video_core/renderer_opengl/gl_shader_gen.h"
-#include "video_core/shader/glsl_decompiler.h"
 #include "video_core/shader/shader_ir.h"
 
 namespace OpenGL::GLShader {
