+++ src/video_core/renderer_opengl/gl_shader_gen.h
@@ -10,7 +10,7 @@
 
 #include "common/common_types.h"
 #include "video_core/engines/shader_bytecode.h"
-#include "video_core/shader/glsl_decompiler.h"
+#include "video_core/renderer_opengl/gl_shader_decompiler.h"
 #include "video_core/shader/shader_ir.h"
 
 namespace OpenGL::GLShader {
